% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_int.R
\name{mle_int}
\alias{mle_int}
\title{Imputation-Based MLE for Interval-Censored Data}
\usage{
mle_int(left, right, dist, true_params = list())
}
\arguments{
\item{left}{Left bounds of censoring intervals}

\item{right}{Right bounds of censoring intervals}

\item{dist}{Distribution name (e.g. \"weibull\", \"loglogistic\", \"EMV\")}

\item{impute}{Imputation method: \"midpoint\", \"random\", etc.}
}
\value{
A list containing estimates, standard errors, and log-likelihood
}
\description{
Estimates distribution parameters using imputed event times.
}
