% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simIC.R
\name{simIC}
\alias{simIC}
\title{Simulate Interval-Censored Survival Data}
\usage{
simIC(
  n = 100,
  dist = "weibull",
  shape = 2,
  scale = 1,
  meanlog = 0,
  sdlog = 1,
  location = 0,
  dist_params = list(),
  interval_width = 1,
  visit_start = 0
)
}
\arguments{
\item{n}{Number of samples.}

\item{dist}{The distribution to use. Options include: "weibull", "exp", "lognormal",
"loglogistic", "logistic", "normal", "gumbel", "gamma", "gompertz".}

\item{shape}{Shape parameter for Weibull, Log-Logistic, etc.}

\item{scale}{Scale parameter for Weibull, Log-Logistic, etc.}

\item{meanlog}{Mean log for Log-Normal distribution.}

\item{sdlog}{Standard deviation for Log-Normal distribution.}

\item{location}{Location parameter for Logistic, Gumbel, etc.}

\item{dist_params}{Optional list of additional distribution parameters.}

\item{interval_width}{Width of the interval for censoring.}

\item{visit_start}{Optional start time for censoring intervals. Default is 0.}
}
\value{
A data frame with the following columns: \code{id}, \code{L}, \code{R}, \code{status}, and \code{true_time}.
}
\description{
This function simulates interval-censored survival data based on the specified distribution.
It supports Weibull, Exponential, Log-Normal, Logistic, and more.
}
\examples{
simIC(n = 15, dist = "weibull", shape = 1.2, scale = 5, interval_width = 4)
simIC(n = 10, dist = "lognormal", meanlog = 3, sdlog = 1, interval_width = 5)

}
